Rename SimpleStorageProvider to KeyedStorageProvider
